Modulariserer dropdown-komponenten
