+++ packages/frontend/components/dropdown/Dropdown.module.css
@@ -1,0 +1,35 @@
+.Dropdown > .navds-popover {
+    padding: 16px 0;
+    border-radius: 4px;
+}
+
+.Dropdown > .navds-popover:focus,
+.Dropdown > .navds-popover:focus-visible {
+    box-shadow: 0 0.05rem 0.25rem 0.125rem rgb(0 0 0 / 8%);
+    border-color: var(--navds-text-focus);
+}
+
+.Button {
+    display: flex;
+    align-items: center;
+    color: var(--navds-color-action-default);
+    font-size: 1rem;
+    font-weight: 600;
+    padding: 8px 12px;
+}
+
+.Button > svg {
+    margin-left: 0.5rem;
+}
+
+.Button:focus-visible {
+    box-shadow: inset var(--navds-shadow-focus);
+}
+
+.List {
+    display: flex;
+    flex-direction: column;
+    list-style: none;
+    padding: 0;
+    background: var(--navds-color-background);
+}
