+++ packages/frontend/components/dropdown/Dropdown.tsx
@@ -1,80 +1,11 @@
-import styled from '@emotion/styled';
 import React, { HTMLAttributes, useState } from 'react';
-
+import classNames from 'classnames';
 import { Collapse, Expand } from '@navikt/ds-icons';
 import { Popover } from '@navikt/ds-react';
 
-import { Button } from '../Button';
+import { Button } from '@components/Button';
 
-const Container = styled.div`
-    > .navds-popover {
-        padding: 16px 0;
-        border-radius: 4px;
-
-        &:focus,
-        &:focus-visible {
-            box-shadow: 0 0.05rem 0.25rem 0.125rem rgb(0 0 0 / 8%);
-            border-color: var(--navds-text-focus);
-        }
-    }
-`;
-
-const ToggleDropdownButton = styled(Button)`
-    display: flex;
-    align-items: center;
-    color: var(--navds-color-action-default);
-    font-size: 1rem;
-    font-weight: 600;
-    padding: 8px 12px;
-
-    > svg {
-        margin-left: 0.5rem;
-    }
-
-    &:focus-visible {
-        box-shadow: inset var(--navds-shadow-focus);
-    }
-`;
-
-export const DropdownButton = styled(Button)`
-    all: unset;
-    height: 32px;
-    min-width: 180px;
-    font-size: 1rem;
-    white-space: nowrap;
-    text-align: left;
-    padding: 0 16px;
-    width: 100%;
-    box-sizing: border-box;
-
-    &:hover,
-    &:focus {
-        background: var(--navds-color-blue-10);
-        color: var(--navds-primary-text);
-        cursor: pointer;
-    }
-
-    &:focus-visible,
-    &:focus {
-        box-shadow: inset 0 0 0 2px var(--navds-text-focus);
-    }
-
-    &:disabled {
-        &,
-        &:hover {
-            background-color: transparent;
-            color: var(--navds-color-text-disabled);
-        }
-    }
-`;
-
-const Liste = styled.ul`
-    display: flex;
-    flex-direction: column;
-    list-style: none;
-    padding: 0;
-    background: var(--navds-color-background);
-`;
+import styles from './Dropdown.module.css';
 
 interface DropdownContextValue {
     lukk: () => void;
@@ -103,11 +34,11 @@ export const Dropdown: React.FC<DropdownProps> = ({ className, onClick, children
     };
 
     return (
-        <Container>
-            <ToggleDropdownButton onClick={onClickWrapper} className={className}>
+        <div className={styles.Dropdown}>
+            <Button className={classNames(styles.Button, className)} onClick={onClickWrapper}>
                 Meny
                 {anchor !== null ? <Collapse /> : <Expand />}
-            </ToggleDropdownButton>
+            </Button>
             <Popover
                 open={anchor !== null}
                 tabIndex={-1}
@@ -117,8 +48,8 @@ export const Dropdown: React.FC<DropdownProps> = ({ className, onClick, children
                 onClose={lukk}
                 offset={0}
             >
-                <Liste>{children}</Liste>
+                <ul className={styles.List}>{children}</ul>
             </Popover>
-        </Container>
+        </div>
     );
 };
