+++ packages/frontend/components/dropdown/index.ts
@@ -1,0 +1,2 @@
+export { DropdownButton } from './DropdownButton';
+export { Dropdown, DropdownContext } from './Dropdown';
